Make disconnecting update contexts. Publish version async.
