+++ src/main/java/no/ndla/taxonomy/service/task/Deleter.java
@@ -9,17 +9,17 @@ package no.ndla.taxonomy.service.task;
 
 import java.util.Optional;
 import no.ndla.taxonomy.domain.DomainEntity;
-import no.ndla.taxonomy.service.NodeService;
+import no.ndla.taxonomy.service.NodeConnectionService;
 
 public class Deleter extends Task<DomainEntity> {
-    private NodeService nodeService;
+    private NodeConnectionService nodeConnectionService;
 
-    public void setNodeService(NodeService nodeService) {
-        this.nodeService = nodeService;
+    public void setNodeConnectionService(NodeConnectionService nodeService) {
+        this.nodeConnectionService = nodeService;
     }
 
     @Override
     protected Optional<DomainEntity> execute() {
-        return nodeService.disconnectAllInvisibleNodes();
+        return nodeConnectionService.disconnectAllInvisibleNodes();
     }
 }
